# internal/web/calculations/api.gen.go
// Package calculations provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package calculations

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// Calculation defines model for Calculation.
type Calculation struct {
	Expression *string `json:"expression,omitempty"`
	Id         *string `json:"id,omitempty"`
	Result     *string `json:"result,omitempty"`
	UserId     *string `json:"user_id,omitempty"`
}

// PostCalculationsJSONRequestBody defines body for PostCalculations for application/json ContentType.
type PostCalculationsJSONRequestBody = Calculation

// PatchCalculationsIdJSONRequestBody defines body for PatchCalculationsId for application/json ContentType.
type PatchCalculationsIdJSONRequestBody = Calculation

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// get all calculations
	// (GET /calculations)
	GetCalculations(ctx echo.Context) error
	// create calculation
	// (POST /calculations)
	PostCalculations(ctx echo.Context) error
	// delete the calculation
	// (DELETE /calculations/{id})
	DeleteCalculationsId(ctx echo.Context, id string) error
	// patch the calculation
	// (PATCH /calculations/{id})
	PatchCalculationsId(ctx echo.Context, id string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetCalculations converts echo context to params.
func (w *ServerInterfaceWrapper) GetCalculations(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetCalculations(ctx)
	return err
}

// PostCalculations converts echo context to params.
func (w *ServerInterfaceWrapper) PostCalculations(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostCalculations(ctx)
	return err
}

// DeleteCalculationsId converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteCalculationsId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteCalculationsId(ctx, id)
	return err
}

// PatchCalculationsId converts echo context to params.
func (w *ServerInterfaceWrapper) PatchCalculationsId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchCalculationsId(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/calculations", wrapper.GetCalculations)
	router.POST(baseURL+"/calculations", wrapper.PostCalculations)
	router.DELETE(baseURL+"/calculations/:id", wrapper.DeleteCalculationsId)
	router.PATCH(baseURL+"/calculations/:id", wrapper.PatchCalculationsId)

}

type GetCalculationsRequestObject struct {
}

type GetCalculationsResponseObject interface {
	VisitGetCalculationsResponse(w http.ResponseWriter) error
}

type GetCalculations200JSONResponse []Calculation

func (response GetCalculations200JSONResponse) VisitGetCalculationsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostCalculationsRequestObject struct {
	Body *PostCalculationsJSONRequestBody
}

type PostCalculationsResponseObject interface {
	VisitPostCalculationsResponse(w http.ResponseWriter) error
}

type PostCalculations201JSONResponse Calculation

func (response PostCalculations201JSONResponse) VisitPostCalculationsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type DeleteCalculationsIdRequestObject struct {
	Id string `json:"id"`
}

type DeleteCalculationsIdResponseObject interface {
	VisitDeleteCalculationsIdResponse(w http.ResponseWriter) error
}

type DeleteCalculationsId204Response struct {
}

func (response DeleteCalculationsId204Response) VisitDeleteCalculationsIdResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PatchCalculationsIdRequestObject struct {
	Id   string `json:"id"`
	Body *PatchCalculationsIdJSONRequestBody
}

type PatchCalculationsIdResponseObject interface {
	VisitPatchCalculationsIdResponse(w http.ResponseWriter) error
}

type PatchCalculationsId200JSONResponse Calculation

func (response PatchCalculationsId200JSONResponse) VisitPatchCalculationsIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// get all calculations
	// (GET /calculations)
	GetCalculations(ctx context.Context, request GetCalculationsRequestObject) (GetCalculationsResponseObject, error)
	// create calculation
	// (POST /calculations)
	PostCalculations(ctx context.Context, request PostCalculationsRequestObject) (PostCalculationsResponseObject, error)
	// delete the calculation
	// (DELETE /calculations/{id})
	DeleteCalculationsId(ctx context.Context, request DeleteCalculationsIdRequestObject) (DeleteCalculationsIdResponseObject, error)
	// patch the calculation
	// (PATCH /calculations/{id})
	PatchCalculationsId(ctx context.Context, request PatchCalculationsIdRequestObject) (PatchCalculationsIdResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetCalculations operation middleware
func (sh *strictHandler) GetCalculations(ctx echo.Context) error {
	var request GetCalculationsRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetCalculations(ctx.Request().Context(), request.(GetCalculationsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetCalculations")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetCalculationsResponseObject); ok {
		return validResponse.VisitGetCalculationsResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PostCalculations operation middleware
func (sh *strictHandler) PostCalculations(ctx echo.Context) error {
	var request PostCalculationsRequestObject

	var body PostCalculationsJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostCalculations(ctx.Request().Context(), request.(PostCalculationsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostCalculations")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostCalculationsResponseObject); ok {
		return validResponse.VisitPostCalculationsResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// DeleteCalculationsId operation middleware
func (sh *strictHandler) DeleteCalculationsId(ctx echo.Context, id string) error {
	var request DeleteCalculationsIdRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteCalculationsId(ctx.Request().Context(), request.(DeleteCalculationsIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteCalculationsId")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteCalculationsIdResponseObject); ok {
		return validResponse.VisitDeleteCalculationsIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PatchCalculationsId operation middleware
func (sh *strictHandler) PatchCalculationsId(ctx echo.Context, id string) error {
	var request PatchCalculationsIdRequestObject

	request.Id = id

	var body PatchCalculationsIdJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PatchCalculationsId(ctx.Request().Context(), request.(PatchCalculationsIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchCalculationsId")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PatchCalculationsIdResponseObject); ok {
		return validResponse.VisitPatchCalculationsIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}