# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: api
  version: 1.0.0

paths:
  /calculations:
    get:
      summary: get all calculations
      tags:
        - calculations
      responses:
        '200':
          description: List of calculations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Calculation'

    post:
      summary: create calculation
      tags:
        - calculations
      requestBody:
        description: take expression
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Calculation'
      responses:
        '201':
          description: create calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calculation'

  /calculations/{id}:
    patch:
      summary: patch the calculation
      tags:
        - calculations
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        description: take new expression
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Calculation'
      responses:
        '200':
          description: updated calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calculation'

    delete:
      summary: delete the calculation
      tags:
        - calculations
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: deleted calculation
  /users:
    get:
      summary: get all users
      tags:
        - users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Users'

    post:
      summary: Create users
      tags:
        - users
      requestBody:
        description: Take info about user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
  /users/{id}:
    get:
      summary: Get user with all his calculations
      tags:
        - users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: user with calculations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'

    patch:
      summary: Update info about user
      tags:
        - users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        description: Take new info about user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '200':
          description: updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
    delete:
      summary: delete the user
      tags:
        - users
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: deleted calculation

components:
  schemas:
    Calculation:
      type: object
      properties:
        id:
          type: string
        expression:
          type: string
        result:
          type: string
        user_id:
          type: string
    Users:
      type: object
      properties:
        id:
          type: string
        Email:
          type: string
        Calculation:
          type: array
          items:
            $ref: '#/components/schemas/Calculation'
        Password:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true